requirements.txt
Django==4.2.3
django-extensions==3.2.3
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0